requirements.txt
# Core MCP
fastmcp>=0.9.0

# Data Analysis
pandas>=2.0.0
openpyxl>=3.1.0

# Visualization
plotly>=5.17.0

# Web & HTTP
httpx>=0.25.0
beautifulsoup4>=4.12.0

# Document Processing
PyPDF2>=3.0.0

# UI
streamlit>=1.28.0
st-social-media-links>=0.1.0

# LangChain & OpenAI
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0

# MCP Client (ajustar según tu librería)
# mcp-use>=0.1.0

# Utilities
python-dotenv>=1.0.0